backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.0
websockets==12.0
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6 